Build Docker image in CI
Adds a workflow to build the included Dockerfile on PRs and merges to
`main`. Additionally, adds a `-version` flag to the binary, which makes
it print out `argv[0]` and the `debug.BuildInfo` ([1]) embedded in the
binary. This is just so that there's some way to run the binary such
that it will exit with status 0 immediately, which lets us verify in CI
that the container was built properly. We also embed the git SHA into a
label on the container image.

[1]: https://pkg.go.dev/runtime/debug#BuildInfo