Add Prometheus metric scrapes
Adds a config option `PROMETHEUS_CONFIG` that instructs the server to
expose a Prometheus metrics scrape endpoint. The value is a JSON
document allowing control over the host and port the metrics handler
will listen on (which may not be the same host/port as the OHTTP
gateway), the path at which scrapes are served and the name of the
metric. The only metric currently served is a histogram of response
times. Its labeling is rather awkward, in an effort to work within the
existing awkward tags on the StatsD metrics, which interpolate HTTP
status into result strings.